PairProblem now uses Points

+++ eclipse/labs/lab4/src/pairsOnAPlane/Point.java
@@ -1,10 +1,10 @@
 package pairsOnAPlane;
 
 public class Point implements Comparable<Point> {
-	public int x;
-	public int y;
+	public double x;
+	public double y;
 	
-	public Point(int x, int y) {
+	public Point(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
@@ -15,6 +15,6 @@ public class Point implements Comparable<Point> {
 	 */
 	@Override
 	public int compareTo(Point that) {
-		return this.x - that.x;
+		return (int) ( this.x - that.x );
 	}
 }
